feat(rag): adiciona pré-carregamento de PDFs via pasta initial_docs
